Changed Some Controllers/ Added Repos / Stored Procedures

+++ Models/Dish.cs
@@ -8,7 +8,7 @@ namespace RestaurantAPI.Models
         [Key]
         public int Dish_ID { get; set; }
         public bool Available { get; set; }
-        public double Price { get; set; }
+        public Decimal Price { get; set; }
         public string Description { get; set; }
         public string Menu_Type { get; set; }
     }
